break out ModelReport as separate component
